+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/) for
 official project tracking. For technical reasons, the module version
 number cannot be expressed using Semantic Versioning.
 
+## 1.3.0 - 2017-11-08
+### Features
+- Test-SemanticVersion has a 'AsErrorRecord' parameter that will return a PowerShell ErrorRecord object
+  if the input string is not a valid Semantic Version string. The ErrorRecord object contains detail on why
+  the string was invalid and can thrown from other functions for validation.
+- Improved feedback messages provided by Test-SemanticVersion.
+
 ## 1.2.0 - 2017-11-01
 ### Features
 - Test-SemanticVersion provides useful feedback messages if a Semantic Version is invalid and the -Verbose switch